Move EosExtensions and EosBranchFile to the util library
We will need them in a library for preparing the update on a volume.
