+++ src/eos-updater-poll-common.h
@@ -22,10 +22,10 @@
 
 #pragma once
 
-#include "eos-updater-branch-file.h"
 #include "eos-updater-data.h"
-#include "eos-updater-extensions.h"
 
+#include "eos-branch-file.h"
+#include "eos-extensions.h"
 #include "eos-refcounted.h"
 
 #include <ostree.h>
